pyproject.toml
[tool.poetry]
name = "djangoreactflash"
version = "0.1.0"
description = ""
authors = ["Liam O'Brien <liam.obrien@atteris.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pytest = "^7.4.1"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
